+++ dangerzone/updater/log_index.py
@@ -1,0 +1,3 @@
+# RELEASE: Bump this value to the log index of the latest signature
+# to ensure the software can't upgrade to container images that predates it.
+LAST_KNOWN_LOG_INDEX = 732661252